cloud_storage: add types common header for clients
When clients for different cloud storage providers will be added, the
common types can be shared via the new header.
